Work on Racer
I did a lot without committing. Very bad.

I imported sprites for testing the racer (thanks Kenney).

I got basic input working for acceleration, reversing, braking, and turning. It also changes the sprite based on which way the car is facing.

The driving is kind of close to what I want with some oddities I need to smooth out. Namely, turning in place adds acceleration and turning doesn't dampen velocity correctly. I don't know much about physics and I've never made a driving game before, so I need to find some good tips for this kind of thing.

I set the resolution to 1280x720. This would be the size for the camera showing the whole race track. I'll design the tilemaps with that in mind; 64x64 tiles to fill a space about 1280x720 (but slightly smaller).

For simplicity, I could just slap a camera on and let everyone see the whole track all at once, or I could opt for the more complex option to make a mini map view of the whole track and give everyone a smaller camera view into the world. "For simplicity" is a good start in a game jam, so just do the macro view first.

All different kinds of racer will use this basic setup for sprite swapping, just with a different sprite folder path. I'll make different selectable racers with just cosmetic differences.

I put a lot of to-dos in the Racer script.